requirements.txt
alembic==1.13.2
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fpdf==1.7.2
future==1.0.0
greenlet==3.1.0
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
ntc_templates==6.0.0
packaging==24.2
paramiko==3.4.1
pillow==11.0.0
pycparser==2.22
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
scp==0.15.0
setuptools==74.1.2
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
textfsm==1.1.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.32.0
Werkzeug==3.0.4
WTForms==3.1.2
